+++ src/indicators/bbands.rs
@@ -1,8 +1,7 @@
+use std::cmp::Ordering;
 use crate::indicators::{Indicator, IndicatorGraphHandler, IndicatorSignalHandler, IndicatorUtilities};
 use crate::types::Signal;
-use chrono::NaiveDateTime;
 use polars::prelude::*;
-use ta::indicators::{BollingerBands, BollingerBandsOutput};
 use ta::Next;
 use crate::utils::extract_new_rows;
 
@@ -11,13 +10,49 @@ const DEFAULT_MULTIPLIER: f64 = 2.0;
 const DEFAULT_THRESHOLD: f64 = 0.8;
 const SOURCE_COL_NAME: &str = "close";
 
+fn calculate_bollinger_bands(df: &DataFrame, column_name: &str, periods: usize, num_std: f64) -> Result<DataFrame, PolarsError> {
+    let mut window_options = RollingOptionsFixedWindow::default();
+    window_options.min_periods = periods;
+    window_options.window_size = periods;
+
+    // Calculate the simple moving average (middle band)
+    let sma = df.column(column_name)?
+        .rolling_mean(RollingOptionsFixedWindow::default())?;
+
+    // Calculate the standard deviation
+    let std_dev = df.column(column_name)?
+        .rolling_std(window_options)?;
+
+    // Calculate upper and lower bands
+    let upper_band = sma.clone() + (std_dev.clone() * num_std);
+    let lower_band = sma.clone() - (&std_dev * num_std);
+
+    let index = df.column("time").unwrap();
+
+    Ok(
+        df!{
+            "time" => index,
+            "lower" => lower_band.unwrap(),
+            "middle" => sma,
+            "upper" => upper_band.unwrap()
+        }?
+    )
+}
+
+
 pub struct BBands {
     period: usize,
     multiplier: f64,
 
-    indicator: BollingerBands,
     history: Option<DataFrame>,
     signals: Option<DataFrame>,
+}
+
+
+struct BollingerBandsOutputNew {
+    pub lower: Option<f64>,
+    pub middle: Option<f64>,
+    pub upper: Option<f64>,
 }
 
 impl BBands {
@@ -26,14 +61,14 @@ impl BBands {
             period,
             multiplier,
             history: None,
-            indicator: BollingerBands::new(period, multiplier).unwrap(),
             signals: None,
         }
     }
 }
 impl IndicatorUtilities for BBands {
     fn restart_indicator(&mut self) {
-        self.indicator = BollingerBands::new(self.period, self.multiplier).unwrap();
+        self.history = None;
+        self.signals = None;
     }
 
 }
@@ -45,84 +80,64 @@ impl Default for BBands {
 }
 
 impl IndicatorGraphHandler for BBands {
-    fn process_existing_candles(&mut self, candles: &DataFrame) {
+    fn process_graph_for_existing(&mut self, candles: &DataFrame) {
         self.restart_indicator();
 
-        let source = candles.column(SOURCE_COL_NAME).unwrap().f64().unwrap();
-        let timestamps = candles.column("time").unwrap().datetime().unwrap();
+        let output = calculate_bollinger_bands(
+            candles,
+            SOURCE_COL_NAME,
+            DEFAULT_PERIOD,
+            DEFAULT_MULTIPLIER
+        ).unwrap();
 
-        let output = source
-            .into_iter()
-            .zip(timestamps.into_iter())
-            .map(|(value, timestamp)| {
-                let output = self.indicator.next(value.unwrap());
-                let timestamp = NaiveDateTime::from_timestamp_millis(timestamp.unwrap()).unwrap();
-                convert_output_to_dataframe(output, timestamp)
-            })
-            .collect::<Vec<DataFrame>>();
-
-        let mut df = output.get(0).unwrap().clone();
-        for i in 1..output.len() {
-            df = df.vstack(output.get(i).unwrap()).unwrap();
-        }
-
-        self.history = Some(df);
+        self.history = Some(output);
     }
 
-    fn process_new_candles(&mut self, row: &DataFrame) {
-        // check for duplicated timestamps
-        let duplicated = extract_new_rows(row, self.history.as_ref().unwrap());
-        assert_eq!(duplicated.height(), 1, "Passed dataframe might have duplicated timestamps.");
+    fn process_graph_for_new_candles(&mut self, candles: &DataFrame) -> Result<(), ()> {
+        // TODO: check that height is greater than window/period
+        assert_ne!(candles.height(), 1, "Dataframe must contain more than one row.");
+
+        // Ensure candles include new data
+        let extracted = extract_new_rows(candles, self.history.as_ref().unwrap());
+        assert_eq!(extracted.height(), 1, "Dataframe does not have new data.");
 
         // check validity of row
-        assert_eq!(row.height(), 1, "Row must be a single row.");
         assert_eq!(
-            row.get_column_names(),
+            candles.get_column_names(),
             ["time", "open", "high", "low", "close", "volume"],
             "Row has incorrect column names"
         );
 
-        // get the candle price from source column
-        let data_point = row
-            .column(SOURCE_COL_NAME)
-            .unwrap()
-            .f64()
-            .unwrap()
-            .get(0)
-            .unwrap();
+        // recalculate bollinger bands for a limited subset
+        let last = candles
+            .tail(Some(self.period));
+        let output = calculate_bollinger_bands(
+            &last,
+            SOURCE_COL_NAME,
+            self.period,
+            self.multiplier
+        ).unwrap();
 
-        // get the timestamp
-        let timestamp = NaiveDateTime::from_timestamp_millis(
-            row.column("time")
-                .unwrap()
-                .datetime()
-                .unwrap()
-                .get(0)
-                .unwrap(),
-        )
-        .unwrap();
-
-        // update the indicator
-        let output = self.indicator.next(data_point);
-
-        // convert the output to a DataFrame
-        let df = convert_output_to_dataframe(output, timestamp);
+        let new_row = output
+            .tail(Some(1));
 
         // update the history
         if let Some(ref mut history) = self.history {
-            *history = history.vstack(&df).unwrap();
+            *history = history.vstack(&new_row).unwrap();
         } else {
-            self.history = Some(df);
+            self.history = Some(new_row);
         }
+
+        Ok(())
     }
 
-    fn get_indicator_history(&self) -> &Option<DataFrame> {
-        &self.history
+    fn get_indicator_history(&self) -> Option<&DataFrame> {
+        self.history.as_ref()
     }
 }
 
 impl IndicatorSignalHandler for BBands {
-    fn process_existing_data(&mut self, candles: &DataFrame) {
+    fn process_signals_for_existing(&mut self, candles: &DataFrame) {
         let candle_col = candles.column(SOURCE_COL_NAME).unwrap();
 
         let timestamps = candles.column("time").unwrap().clone();
@@ -142,7 +157,7 @@ impl IndicatorSignalHandler for BBands {
                 };
 
                 let signal = calculate_signal(series, close_price, DEFAULT_THRESHOLD);
-                signal.into()
+                signal.unwrap_or(Signal::Hold).into()
             })
             .collect::<Vec<i32>>();
 
@@ -155,29 +170,33 @@ impl IndicatorSignalHandler for BBands {
         );
     }
 
-    fn process_new_data(&mut self, row: &DataFrame) {
+    fn process_signals_for_new_candles(&mut self, candles: &DataFrame) -> Result<(), ()> {
+
+
+       // TODO: this has to be updated!
+
+
         let graph_row = extract_new_rows(
             self.history.as_ref().unwrap(),
             self.signals.as_ref().unwrap(),
         );
+        assert_eq!(graph_row.height(), 1, "Indicator graph is too ahead of signals. Call bootstrapping again.");
 
-        let duplicated = extract_new_rows(row, self.signals.as_ref().unwrap());
-        assert_eq!(duplicated.height(), 1, "Passed dataframe might have duplicated timestamps.");
+        let new_row = extract_new_rows(candles, self.signals.as_ref().unwrap());
+        assert_eq!(new_row.height(), 1, "Passed dataframe might have duplicated timestamps.");
 
-        assert_eq!(graph_row.height(), 1, "Graph row must be a single row.");
-        assert_eq!(row.height(), 1, "Candle row must be a single row.");
         assert_eq!(
             graph_row.column("time").unwrap().datetime().unwrap().get(0),
-            row
+            new_row
                 .column("time")
                 .unwrap()
                 .datetime()
                 .unwrap()
                 .get(0),
-            "Graph row and candle row must have the same timestamp"
+            "Graph row and new candle row must have the same timestamp"
         );
 
-        let candle_price = row
+        let candle_price = new_row
             .column(SOURCE_COL_NAME)
             .unwrap()
             .f64()
@@ -190,14 +209,15 @@ impl IndicatorSignalHandler for BBands {
             .into_struct("row")
             .into_iter()
             .map(|series| {
-                let signal = calculate_signal(series, candle_price, DEFAULT_THRESHOLD);
+                let signal = calculate_signal(series, candle_price, DEFAULT_THRESHOLD)
+                    .unwrap_or(Signal::Hold);
                 signal.into()
             })
             .collect::<Vec<i32>>();
 
         // update the signals
         let df = df!(
-            "time" => row.column("time").unwrap(),
+            "time" => new_row.column("time").unwrap(),
             "signal" => graph_row
         )
         .unwrap();
@@ -207,88 +227,22 @@ impl IndicatorSignalHandler for BBands {
         } else {
             self.signals = Some(df);
         }
+
+        Ok(())
     }
 
-    fn get_signal_history(&self) -> &Option<DataFrame> {
-        &self.signals
+    fn get_signal_history(&self) -> Option<&DataFrame> {
+        self.signals.as_ref()
     }
 }
 
 impl Indicator for BBands {}
 
-/// Convert indicator output to a DataFrame
-///
-/// This is used to convert the indicator output to a DataFrame with a single row
-/// used to build the history DataFrame.
-///
-/// # Arguments
-/// * `output` - The indicator output (`ta` specific)
-/// * `timestamp` - The timestamp of the candle
-///
-/// # Returns
-/// A DataFrame with a single row to use as a history row
-fn convert_output_to_dataframe(
-    output: BollingerBandsOutput,
-    timestamp: NaiveDateTime,
-) -> DataFrame {
-    let lower = output.lower;
-    let middle = output.average;
-    let upper = output.upper;
-
-    let df = df!(
-            "time" => &[timestamp],
-            "lower" => &[lower],
-            "middle" => &[middle],
-            "upper" => &[upper],
-        )
-        .unwrap();
-
-    df
-}
-
-/// Unwrap a row from the indicator graph into a `BollingerBandsOutput`
-///
-/// The expected series index values are as follows:
-/// * 0: the timestamp
-/// * 1: the lower bound
-/// * 2: the middle/average bound
-/// * 3: the upper bound
-///
-/// # Arguments
-/// * `series` - The row from the indicator graph.
-///
-/// # Returns
-/// A `BollingerBandsOutput` struct
-fn unwrap_output_series(series: &[AnyValue]) -> BollingerBandsOutput {
-    assert_eq!(series.len(), 4, "Series must have 4 values");
-
-    let lower = series.get(1).unwrap();
-    let middle = series.get(2).unwrap();
-    let upper = series.get(3).unwrap();
-
-    let lower = if let AnyValue::Float64(inner) = lower {
-        *inner
-    } else {
-        panic!("Could not get lower value from time-series chart")
-    };
-
-    let middle = if let AnyValue::Float64(inner) = middle {
-        *inner
-    } else {
-        panic!("Could not get middle value from time-series chart")
-    };
-
-    let upper = if let AnyValue::Float64(inner) = upper {
-        *inner
-    } else {
-        panic!("Could not get upper value from time-series chart")
-    };
-
-    BollingerBandsOutput {
-        lower,
-        average: middle,
-        upper,
-    }
+#[derive(Debug)]
+enum BBandExtractionError {
+    InvalidSeriesLength,
+    InvalidDataType,
+    MissingValue,
 }
 
 /// Calculate signal from indicator and close price
@@ -297,36 +251,49 @@ fn unwrap_output_series(series: &[AnyValue]) -> BollingerBandsOutput {
 /// Bollinger Bands.
 ///
 /// # Arguments
-/// * `series` - The indicator series. The values are fed to the `unwrap_output_series()` function
+/// * `series` - The indicator series.
 /// * `candle_price` - The current candle price.
 /// * `threshold` - The threshold to use when calculating the signal. This is expected to be a percentage.
 ///     The higher the value, the more closely the candle price must be to the bounds of the Bollinger Bands
 ///
 /// # Returns
 /// A `Signal` enum
-fn calculate_signal(series: &[AnyValue], candle_price: f64, threshold: f64) -> Signal {
-    let output = unwrap_output_series(series);
-
-    let buy_padding = (output.average - output.lower) * threshold;
-    let buy_threshold = output.average - buy_padding;
-
-    let sell_padding = (output.upper - output.average) * threshold;
-    let sell_threshold = output.average + sell_padding;
-
-    if candle_price <= buy_threshold {
-        Signal::Buy
-    } else if candle_price >= sell_threshold {
-        Signal::Sell
-    } else {
-        Signal::Hold
+fn calculate_signal(series: &[AnyValue], candle_price: f64, threshold: f64) -> Result<Signal, BBandExtractionError> {
+    if series.len() != 4 {
+        return Err(BBandExtractionError::InvalidSeriesLength);
     }
+
+    let extract_float = |index: usize| -> Result<f64, BBandExtractionError> {
+        match series.get(index) {
+            Some(AnyValue::Float64(value)) => Ok(*value),
+            Some(_) => Err(BBandExtractionError::InvalidDataType),
+            None => Err(BBandExtractionError::MissingValue),
+        }
+    };
+
+    let lower = extract_float(1)?;
+    let middle = extract_float(2)?;
+    let upper = extract_float(3)?;
+
+    let buy_threshold = middle - (middle - lower) * threshold;
+    let sell_threshold = middle + (upper - middle) * threshold;
+
+    // TODO: add flag to hold if indicators are equal to thresholds
+    Ok(match candle_price.partial_cmp(&buy_threshold) {
+        Some(Ordering::Less) => Signal::Buy,
+        Some(Ordering::Equal) => Signal::Buy,
+        _ => match candle_price.partial_cmp(&sell_threshold) {
+            Some(Ordering::Greater) => Signal::Sell,
+            Some(Ordering::Equal) => Signal::Sell,
+            _ => Signal::Hold,
+        },
+    })
 }
 
 #[cfg(test)]
 mod tests {
     use chrono::Utc;
     use polars::prelude::*;
-    use ta::Period;
 
     use crate::indicators::bbands::{DEFAULT_MULTIPLIER, DEFAULT_PERIOD};
     use crate::indicators::{IndicatorGraphHandler, IndicatorSignalHandler, IndicatorUtilities};
@@ -339,9 +306,6 @@ mod tests {
         assert_eq!(bb.multiplier, 4.0);
         assert_eq!(bb.history, None);
         assert_eq!(bb.signals, None);
-
-        assert_eq!(bb.indicator.period(), 15);
-        assert_eq!(bb.indicator.multiplier(), 4.0);
     }
 
     #[test]
@@ -351,29 +315,38 @@ mod tests {
         assert_eq!(bb.multiplier, 2.0);
         assert_eq!(bb.history, None);
         assert_eq!(bb.signals, None);
-
-        assert_eq!(bb.indicator.period(), DEFAULT_PERIOD);
-        assert_eq!(bb.indicator.multiplier(), DEFAULT_MULTIPLIER);
     }
 
     #[test]
     fn test_restart_indicator() {
         let mut bb = super::BBands::new(15, 4.0);
+        bb.history = Some(df!{
+            "time" => &[Utc::now().naive_utc()],
+            "lower" => &[1.0],
+            "middle" => &[2.0],
+            "upper" => &[3.0],
+        }.unwrap());
+
+        bb.signals = Some(df!{
+            "time" => &[Utc::now().naive_utc()],
+            "signal" => &[1],
+        }.unwrap());
+
         bb.restart_indicator();
 
-        assert_eq!(bb.indicator.period(), 15);
-        assert_eq!(bb.indicator.multiplier(), 4.0);
+        assert!(bb.history.is_none());
+        assert!(bb.signals.is_none());
     }
 
     #[test]
     fn test_process_existing_candles() {
-        let mut bb = super::BBands::new(4, 2.0);
+        let mut bb = super::BBands::new(DEFAULT_PERIOD, DEFAULT_MULTIPLIER);
 
         // set the candles
         let time = Utc::now().naive_utc();
         let date_range = date_range(
             "time",
-            time - chrono::Duration::minutes(5),
+            time - chrono::Duration::minutes(25),
             time,
             Duration::parse("1m"),
             ClosedWindow::Left,
@@ -383,57 +356,58 @@ mod tests {
         .unwrap();
         let candles = df!(
             "time" => date_range.clone(),
-            "open" => &[1, 2, 3, 4, 5],
-            "high" => &[1, 2, 3, 4, 5],
-            "low" => &[1, 2, 3, 4, 5],
-            "close" => &[1.0, 2.0, 3.0, 4.0, 5.0],
-            "volume" => &[1, 2, 3, 4, 5],
+            "open" => &[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25],
+            "high" => &[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25],
+            "low" => &[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25],
+            "close" => &[
+                1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0,
+                14.0, 15.0, 16.0, 17.0, 18.0, 19.0, 20.0, 21.0, 22.0, 23.0, 24.0, 25.0],
+            "volume" => &[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25],
         )
         .unwrap();
 
-        bb.process_existing_candles(&candles);
+        bb.process_graph_for_existing(&candles);
 
         let history = bb.history.as_ref().unwrap();
 
-        assert_eq!(history.shape(), (5, 4));
+        assert_eq!(history.shape(), (25, 4));
 
-        for i in 0..6 {
+        // ensure that history has correct timestamp
+        for i in 0..date_range.len() {
             assert_eq!(
                 history.column("time").unwrap().datetime().unwrap().get(i),
                 date_range.get(i)
             );
+        };
+
+        // ensure that upper and lower bounds have null values
+        for i in 0..19 {
+            assert_eq!(
+                history.column("lower").unwrap().f64().unwrap().get(i),
+                None
+            );
+            assert_eq!(
+                history.column("upper").unwrap().f64().unwrap().get(i),
+                None
+            );
         }
 
-        assert_eq!(
-            history.column("lower").unwrap().f64().unwrap().get(0),
-            Some(1.0)
-        );
-        assert_eq!(
-            history.column("lower").unwrap().f64().unwrap().get(1),
-            Some(0.5)
-        );
+        // ensure that upper/lower bounds have generally correct values
+        // ensure that middle band is correct
+        for i in 19..25 {
+            // check for general ranges
+            assert!(history.column("lower").unwrap().f64().unwrap().get(i).unwrap() < i as f64 - 2.0);
+            assert!(history.column("upper").unwrap().f64().unwrap().get(i).unwrap() > i as f64 + 2.0);
 
-        assert_eq!(
-            history.column("middle").unwrap().f64().unwrap().get(0),
-            Some(1.0)
-        );
-        assert_eq!(
-            history.column("middle").unwrap().f64().unwrap().get(1),
-            Some(1.5)
-        );
-
-        assert_eq!(
-            history.column("upper").unwrap().f64().unwrap().get(0),
-            Some(1.0)
-        );
-        assert_eq!(
-            history.column("upper").unwrap().f64().unwrap().get(1),
-            Some(2.5)
-        );
+            assert_eq!(
+                history.column("middle").unwrap().f64().unwrap().get(i).unwrap(),
+                i as f64
+            );
+        }
     }
 
     #[test]
-    fn test_process_new_candles() {
+    fn test_process_graph_for_new_candles() {
         // create candles
         let time = Utc::now().naive_utc();
         let date_range = date_range(
@@ -459,7 +433,7 @@ mod tests {
         // create indicator and run `process_existing_candles()`
         let mut bb = super::BBands::new(4, 2.0);
 
-        bb.process_existing_candles(&candles);
+        bb.process_graph_for_existing(&candles);
 
         // assert that the history aligns with candle dimensions
         assert_eq!(bb.history.as_ref().unwrap().height(), 5);
@@ -474,7 +448,10 @@ mod tests {
             "volume" => &[6],
         )
         .unwrap();
-        bb.process_new_candles(&new_row);
+        let new_data = candles.vstack(&new_row).unwrap();
+        let result = bb.process_graph_for_new_candles(&new_data);
+
+        assert!(result.is_ok());
 
         // assert that `history` has been updated with new row
         let history = bb.history.as_ref().unwrap();
@@ -482,12 +459,12 @@ mod tests {
         assert_eq!(history.height(), 6);
         assert_eq!(
             history.column("time").unwrap().datetime().unwrap().get(5),
-            Some(time.timestamp_millis())
+            Some(time.and_utc().timestamp_millis())
         );
     }
 
     #[test]
-    fn test_process_existing_data() {
+    fn test_process_signals_for_existing() {
         // create candles
         let time = Utc::now().naive_utc();
         let date_range = date_range(
@@ -543,7 +520,7 @@ mod tests {
         let mut bb = super::BBands::new(4, 2.0);
         bb.history = Some(history);
 
-        bb.process_existing_data(&candles);
+        bb.process_signals_for_existing(&candles);
 
         bb.signals
             .unwrap()
@@ -563,7 +540,7 @@ mod tests {
     }
 
     #[test]
-    fn test_process_new_data() {
+    fn test_process_signals_signals_for_new_candles() {
         let time = Utc::now().naive_utc();
         let date_range = date_range(
             "time",
@@ -611,18 +588,28 @@ mod tests {
         let history = bb.history.as_ref().unwrap().vstack(&new_history_row).unwrap();
         bb.history = Some(history);
 
+        let old_candles = df!(
+            "time" => date_range.clone(),
+            "open" => &[1, 2, 3, 4, 5],
+            "high" => &[1, 2, 3, 4, 5],
+            "low" => &[1, 2, 3, 4, 5],
+            "close" => &[0.9, 1.1, 1.3, 1.7, 1.9],
+            "volume" => &[1, 2, 3, 4, 5])
+            .unwrap();
         let new_row = df!(
             "time" => &[time],
             "open" => &[6],
             "high" => &[6],
             "low" => &[6],
             "close" => &[6.0],
-            "volume" => &[6],
-        )
-        .unwrap();
+            "volume" => &[6])
+            .unwrap();
+
+        let candles = old_candles.vstack(&new_row).unwrap();
 
         // call process_new_data() and assert that signals have been updated
-        bb.process_new_data(&new_row);
+        let result = bb.process_signals_for_new_candles(&candles);
+        assert!(result.is_ok());
 
         assert_eq!(bb.signals.as_ref().unwrap().height(), 6);
         assert_eq!(