+++ src/portfolio/tracked.rs
@@ -64,7 +64,7 @@ impl TrackedValue {
         // find last row
         let last_row = self
             .0
-            .sort(["timestamp"], false, true)
+            .sort(["timestamp"], SortMultipleOptions::default().with_nulls_last_multi([false, true]))
             .unwrap()
             .tail(Some(1));